disks: disk_usage: partial revert d5d1ac

+++ src/disks/mod.rs
@@ -30,6 +30,7 @@ pub struct IoBlock {
 }
 
 /// Return the (total, free) space of a Disk from it's path (mount_point).
+#[allow(clippy::unnecessary_cast)]
 pub fn disk_usage<P>(path: P) -> Result<(u64, u64), Error>
 where
     P: AsRef<[u8]>,
@@ -41,8 +42,10 @@ where
     }
 
     let statvfs = unsafe { statvfs.assume_init() };
-    let total = statvfs.f_blocks * statvfs.f_frsize;
-    let free = statvfs.f_bavail * statvfs.f_frsize;
+    // The cast here is needed for macOS. As it doesn't hurt on Linux,
+    // don't do some cfg(target = XXX) machinery to workaround clippy.
+    let total = statvfs.f_blocks as u64 * statvfs.f_frsize as u64;
+    let free = statvfs.f_bavail as u64 * statvfs.f_frsize as u64;
 
     Ok((total, free))
 }
